Fix the package requirements for Fedora
I'm not sure why MariaDB-client was added here, because it's not a
valid package on Fedora. However, it *is* a package name on
openSUSE so I moved it to it's own tab. Additionally,
mariadb-client is the name of the required package on Fedora - not
libmariadb-devel (that's also openSUSE.)

The name of the development-tools group is also wrong, it's
development-tools and newer DNF doesn't support groupinstall.